+++ src/DetailsView/reports/components/report-sections/full-rule-header.tsx
