+++ src/tests/unit/tests/DetailsView/reports/components/report-sections/full-rule-detail.test.tsx
@@ -2,13 +2,12 @@
 // Licensed under the MIT License.
 import { shallow } from 'enzyme';
 import * as React from 'react';
-
 import { allInstanceOutcomeTypes } from '../../../../../../../DetailsView/reports/components/instance-outcome-type';
 import {
     FullRuleHeader,
     FullRuleHeaderDeps,
     FullRuleHeaderProps,
-} from '../../../../../../../DetailsView/reports/components/report-sections/full-rule-detail';
+} from '../../../../../../../DetailsView/reports/components/report-sections/full-rule-header';
 import { RuleResult } from '../../../../../../../scanner/iruleresults';
 
 describe('FullRuleHeader', () => {
